fix: merge to view for create project allow functionality for both buttons
