--- dev-client/src/components/projects/ProjectListView/ProjectsEmptyView.tsx
@@ -1,24 +1,0 @@
-import {Box, Heading, Link, Text, VStack} from 'native-base';
-import {useTranslation} from 'react-i18next';
-import {IconButton} from '../../common/Icons';
-import AddButton from '../../common/AddButton';
-
-/**
- * Component for displaying info when a user doesn't have any projects
- */
-export default function ProjectsEmptyView() {
-  const {t} = useTranslation();
-  return (
-    <VStack m="3" mt="5" flexGrow={1}>
-      <Heading size="sm">{t('projects.none.header')}</Heading>
-      <Text>{t('projects.none.info')}</Text>
-      <Link _text={{color: 'primary.main'}} alignItems="center" mb="4">
-        <IconButton name="open-in-new" _icon={{color: 'action.active'}} />
-        {t('projects.learn_more')}
-      </Link>
-      <Box alignItems="flex-start">
-        <AddButton text={t('projects.create_button')} />
-      </Box>
-    </VStack>
-  );
-}
